remove the mock-types crate
