+++ toolkit/primitives/authority-selection-inherents/Cargo.toml
@@ -31,7 +31,6 @@ sp-consensus-slots = { workspace = true }
 [dev-dependencies]
 hex = { workspace = true }
 hex-literal = { workspace = true }
-mock-types = { workspace = true, features = ["std"] }
 num-bigint = { workspace = true }
 num-traits = { workspace = true }
 tokio = { workspace = true }
