+++ toolkit/primitives/authority-selection-inherents/src/tests.rs
@@ -3,10 +3,13 @@ use crate::filter_invalid_candidates::RegisterValidatorSignedMessage;
 use crate::select_authorities::select_authorities;
 use hex_literal::hex;
 use num_bigint::BigInt;
+use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
 use plutus::Datum::{ByteStringDatum, ConstructorDatum, IntegerDatum};
 use plutus::ToDatum;
+use scale_info::TypeInfo;
+use serde::{Deserialize, Serialize};
 use sidechain_domain::*;
-use sp_core::{ecdsa, ed25519, ConstU32, Pair};
+use sp_core::{ecdsa, ed25519, sr25519, ConstU32, Pair};
 use sp_runtime::traits::Zero;
 use std::collections::HashMap;
 
@@ -71,7 +74,40 @@ impl TryFrom<SidechainPublicKey> for AccountId {
 	}
 }
 
-pub type AccountKeys = mock_types::session_keys::SessionKeys;
+#[derive(
+	Clone,
+	Debug,
+	PartialEq,
+	Eq,
+	PartialOrd,
+	Ord,
+	Encode,
+	Decode,
+	TypeInfo,
+	MaxEncodedLen,
+	Serialize,
+	Deserialize,
+)]
+pub struct AccountKeys {
+	pub aura: [u8; 32],
+	pub grandpa: [u8; 32],
+}
+
+impl From<(sr25519::Public, ed25519::Public)> for AccountKeys {
+	fn from((aura, grandpa): (sr25519::Public, ed25519::Public)) -> Self {
+		Self { aura: aura.0, grandpa: grandpa.0 }
+	}
+}
+
+impl AccountKeys {
+	pub fn from_seed(seed: &str) -> AccountKeys {
+		let mut aura = format!("aura-{seed}").into_bytes();
+		aura.resize(32, 0);
+		let mut grandpa = format!("grandpa-{seed}").into_bytes();
+		grandpa.resize(32, 0);
+		AccountKeys { aura: aura.try_into().unwrap(), grandpa: grandpa.try_into().unwrap() }
+	}
+}
 
 #[derive(Clone)]
 pub(crate) struct MockValidator {
