--- toolkit/primitives/mock-types/src/lib.rs
@@ -1,3 +1,0 @@
-#![cfg_attr(not(feature = "std"), no_std)]
-
-pub mod session_keys;
